Modularize chat block view.ts into focused modules
